// src/app/products/model/page.ts
export interface Page<T> {
    content: Array<T>;
    pageable: Pageable;
    last: boolean;
    totalElements: number;
    totalPages: number;
    size: number;
    number: number;
    sort: Sort;
    first: boolean;
    numberOfElements: number;
    empty: boolean;
  };

  export interface Pageable {
    sort: Sort;
    offset: number;
    pageSize: number;
    pageNumber: number;
    unpaged: boolean;
    paged: boolean;
  };

  export interface Sort {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };